move Contracts Settings to the Main Settings
