+++ src/pages/Manage/Settings/Main/index.js
@@ -1,6 +1,7 @@
 import * as s from "../../../../styles/global";
 import IPFS from './IPFS';
 import Networks from './Networks';
+import Contracts from './Contracts';
 
 
 export default function Main() {
@@ -10,6 +11,8 @@ export default function Main() {
       <IPFS />
       <s.SpacerSmall />
       <Networks />
+      <s.SpacerSmall />
+      <Contracts />
     </>
   )
 }
